+++ src/plastimatch/base/rpl_volume.cxx
@@ -285,7 +285,7 @@ static
 void
 proton_dose_ray_trace_callback (
     void *callback_data, 
-    int vox_index, 
+    size_t vox_index, 
     double vox_len, 
     float vox_value
 )
